\name{SpatialNP-package}
\alias{SpatialNP-package}
\alias{SpatialNP}
\docType{package}
\title{
\packageTitle{SpatialNP}
}
\description{
\packageDescription{SpatialNP}
}
\details{
\packageDESCRIPTION{SpatialNP}

There are three functions for inference,
\code{\link{sr.loc.test}}, \code{\link{sr.indep.test}} and
\code{\link{sr.sphere.test}}, for location, independence and
sphericity tests. The so called \link[=rank.shape]{inner} and
\link[=SCov]{outer} standardization matrices are also available as
well as the actual \link[=spatial.rank]{sign and rank} score
functions, together with a utility function \code{\link{to.shape}}.  }


\author{
\packageAuthor{SpatialNP}

Maintainer: \packageMaintainer{SpatialNP}
}

\keyword{ package }
\seealso{
package \pkg{\link[ICSNP:ICSNP-package]{ICSNP}} \cr
package \pkg{\link[MNM:MNM-package]{MNM}}}


